// admin/src/app/_barber/model/barber.ts
export class Barber {

    name: string;
    surname: string;
    company: string;
    phone: string;
    email: string;
    password: string;
    ip: string;
    active: boolean;
    document: string;
    taxPlace: string;
    taxNo: string;
    tcId: string;
    city: string;
    state: string;
    openAddress:string;


}